audio-scribe-project-LLM/requirements.txt
azure-storage-blob==12.20.0
fastapi==0.111.0
numpy==1.26
python-dotenv==1.0.1
torch==2.2.2
transformers==4.41.2
uvicorn==0.30.1
pydantic==2.8.2
langchain==0.2.11
langchain_core==0.2.25
langchain_community==0.2.10
langchain-openai==0.1.19
langchain_chroma==0.1.2
pymongo
